Code cleanup

+++ include/SampleClass.h
@@ -1,8 +1,17 @@
+//******************************************************************************
+// File name:   SampleClass.h
+// Author:      
+// Date:        
+// Class:       
+// Assignment:  
+// Purpose:     
+// Hours:       
+//******************************************************************************
 #pragma once
 
 #include <iostream>
 
 class SampleClass {
 public:
-  unsigned int factorial (unsigned int uInt);
+  size_t factorial (size_t uInt);
 };
